Release task now always publishes to staging
Production builds should never be published directly from a local
machine. Instead, the release command will now publish to a
commit-specific staging URL, so you use it to publish a release
candidate, and then when those builds have been verified, you need to
copy the RC builds from on the staging location on s3 to the production
folder.


Former-commit-id: 1e2c953f16ab6234d8be7a85be17205d714a77ca