Add a new GET script languages API endpoint
To provide additional language options for scripted fields, Kibana needs
to know what languages are available for inline scripting in
Elasticsearch. This new Kibana API grabs up to date information from
Elasticsearch and provides it in an easy to digest format for the UI.

The response body from the API is a simple JSON array with a list of
languages that are enabled for inline scripting. The API also filters
out languages that don't make sense in scripted fields, like 'mustache'.

An example request might look like this:

`GET /api/kibana/scripts/languages`

200 OK
`['expression', 'painless']`


Former-commit-id: 6c8ec66b3ab43875cc1e9c153f45941416700f45