Fix filtering from visualizations
The ability to create filters by clicking on visualizations was broken
between 5.0 alpha3 and alpha4. The root cause was the addition of a new
parameter being sent to all visualization event handlers, which was
added to support persistence of tilemap position and zoom level. The
filter_bar_click_handler (which enables the vis filtering behavior)
already expected a different kind of second parameter, so the addition
of this new parameter threw off the click handler's internal logic.

It turns out that the new parameter is unnecessary, because the object
it's passing is already available on the event object. So to fix this,
I removed the new parameter from the emitter and updated the tilemap
handlers to grab the object off of the event.

Fixes: https://github.com/elastic/kibana/issues/7501
Related: https://github.com/elastic/kibana/pull/6835 (added the new
param)

Former-commit-id: 20cb910f563dfd0f7b30d63317ced57051e03230