Configurable headers for all elasticsearch requests
A new server-side configuration, elasticsearch.customHeaders, allows
people to configure any number of custom headers that will get sent
along to all requests to Elasticsearch that are made via the proxy or
exposed client.

This allows for advanced architectures that do things such as dynamic
routing based on install-specific headers.


Former-commit-id: d00d177d012fc8fae610f5f5e998d8b8cae5ca55