Deprecate ascending sort for terms aggregations
Elasticsearch is deprecating the ability to use ascending sort for terms
aggregations, so we render a deprecation notice whenever a user views a
visualization that uses a terms aggregation with ascending sort by
count. We only render this error message once per route so it does not
spam the users constantly when they have things like auto-refresh
enabled on a dashboard.


Former-commit-id: 60a7b75ad257baab5e048eb84e8c9fb1d3e0b55c