Backport PR #8212 ---------
**Commit 1:**
[state/unhashUrl] use encode-uri-query to generate cleanly encoded urls

By default, the url.format function stringifies the query object and encodes each parameter with `encodeURIComponent()`. This is effective, but does not match the native browser behavior, which is to only encode the non-`pchar` characters as defined by [rfc3986][1]. Because of this, angular's `$location.search()` function uses a function internally called `encodeUriQuery()` to mimic the browser behavior before updating `location.href`.

To bring the same funtionality to the `unhashUrl()` function, the `encodeUriQuery()` was extracted into it's own package and used as the escape handler for `querystring.stringify()`

[1]: https://www.ietf.org/rfc/rfc3986.txt

* Original sha: 41d439946b0a510daa55e5695238ba4e6bbe58bf [formerly d9dac3ffdcd5f7d963784e39e6c35b36b2d28cb6]
* Authored by spalger <spalger@users.noreply.github.com> on 2016-09-09T22:39:57Z


Former-commit-id: c257bbb74d6584d1640def883e011276e6da088b