Backport #8355
[kbnDevToolsApp] allow extending the top nav

The kbn-dev-tools-app directive renders a navbar above each devtool, giving it breadcrumbs and links to the other devtools. This collides with the toolbars used in devtools like console and causes two mostly-empty toolbars to be stacked.

To fix this, we are merging the navbar from kbn-dev-tools-app and the toolbar provided by the devtools into a single kbn-top-nav toolbar. Console already used the kbn-top-nav directive for it's navbar, so kbn-dev-tools-app just needed to be extended to consume and render the kbn-top-nav configuration (a controller in this instance).

(cherry picked from commit b0b9026f06da214e23df8fc4b636fa9532f2fbcc [formerly 5bff9daa94b1acd62ed14209f04d36169e9788af])


Former-commit-id: 410f61f074d3ee1d6cac17c165a1f0f13b9a1cb2