Backport PR #8336 ---------
**Commit 1:**
make the legend background color transparent so it looks good in both dark and light themes

* Original sha: 3e8d058da1e96cb51ccb9765179af5c8bb997196 [formerly 6e94ff5e4516657c99b8b42b0a739b30a5c3d771]
* Authored by Stacey-Gammon <gammon@elastic.co> on 2016-09-19T17:36:57Z

**Commit 2:**
Remove total transparency fix

If the legend has a totally transparent background it’s difficult to
read when the chart lines lie under it.  This approaches fixes the
coloring but leaves the opacity in tact.

* Original sha: c4ad9c1d80a77c04b030ac9de9c94c5443ce29dd [formerly a62b1f0576be16c2f7166576fcc5f71c6d3c6625]
* Authored by Stacey-Gammon <gammon@elastic.co> on 2016-09-19T19:01:21Z


Former-commit-id: f108b11560479b19ccfbb1e3a8f67ba78e0ded64