Backport PR #8513 ---------
**Commit 1:**
timelion: Mark third party services as experimental

We don't have any mechanism in place right now to verify that the third
party backends that ship with Timelion are continuing to work as
expected, and in the event that any service introduces a backwards
compatibility break or becomes unavailable, we have no plan for how
users can workaround the issue.

This is essentially the same situation we found ourselves in with our
third party tile map provider, so we should at least come up with
solutions for this before removing the experimental labels.

* Original sha: 423a8468c940ecaa7721aece5f6db1e628bd13c1 [formerly 7af23789e5ec06d3fbdf2b7aa6e79f2b0c38d6a8]
* Authored by Court Ewing <court@epixa.com> on 2016-10-02T15:01:33Z


Former-commit-id: 4c1fb83d2aa994db91efbfd94887884e318cd1e6