Backport PR #8538 ---------
**Commit 1:**
[agg_response/point_series] always assign series labels to points

Since forever, the vislib has been compensating for a behavior/bug in the agg_response module where "series" labels were not assigned to points if there wasn't a "series" aggregation or multiple y-axis. To compensate for this, the vislib was assigning the yAxisLabel to the series, which had the same effect, but now that the legend is rendered in it's own tick, followed by the visualization, and that behavior was being executed during the vislib render, the legend has been flashing an empty row for a while (this was worsened by the fact that vislib rendering did not trigger a digest cycle).

These changes remove the workaround code from the vislib, and update the agg_response to always send the y-axis label as the series label when there is no series aggregation. Behavior with multiple-y-axis and multiple series aggs should be preserved.

* Original sha: e318284fb379c099ddfc424d41ce58fdef538b26 [formerly 7460c0ac2d038dae5bcfe5ced735be0b1df6ccf0]
* Authored by spalger <email@spalger.com> on 2016-10-04T18:40:41Z

**Commit 2:**
[agg_response/point_series] update tests to reflect changes in 19773a6

* Original sha: 2cbcb62fcaff10f1f2aa260aef5a01e66ba4e06c [formerly f681d6cf865efaa99d8cf3d31334089f8dd2dcd6]
* Authored by spalger <email@spalger.com> on 2016-10-04T19:22:55Z

**Commit 3:**
[agg_response/point_series] apply PR feedback

* Original sha: f207c7b53798b92abab054c0c65982ae9c92c821 [formerly a119577af83d5e0dc44bf828ed05ce56d93a7830]
* Authored by spalger <email@spalger.com> on 2016-10-04T20:41:29Z

**Commit 4:**
[agg_response/point_series] use no-digest promises in tests

* Original sha: 7cf4efdeef5c10b408cb4da3ef1a4f67b72c5f00 [formerly 8d370826c4b73ea25f7d622decca66823bb3a821]
* Authored by spalger <email@spalger.com> on 2016-10-04T20:45:02Z


Former-commit-id: a0fb59f54d0e5e6c7cc0cfcd40426c904485b4e1