Backport PR #7990 ---------
**Commit 1:**
Add indicies and type information for conflict field

Fixes #7661

Added indicies and type information when hover over field with conflict type in
index patterns page.

* Original sha: 86e332ce3cd77476212f43b8095902191c008188 [formerly f842f65fafa58b1ca2d1b90a8862033592862401]
* Authored by Martin Hickey <martin.hickey@ie.ibm.com> on 2016-08-12T12:35:01Z

**Commit 2:**
Update to display the conflict information on the edit field page

Fixes #7661

Moved content to the edit field page from the tooltip for better scalability when large data.
The conflict description data is now only stored for conflicting fields.

* Original sha: ae234ade6bb2ab2c37268f347d52d31c3925f66b [formerly 0af13831d5f62be15da3cf807bf35bb7c17bf4ba]
* Authored by Martin Hickey <martin.hickey@ie.ibm.com> on 2016-09-08T16:15:19Z

**Commit 3:**
Update the conflict data to show all index names per type

Fixes #7661

Feedback from review is that it would be better to display a list of indices
for each field type as would be easier to distinguish the outliers in this
case.

* Original sha: 53eb471bfa29ee75c4d9f39376496a1d7bc167a6 [formerly ee698c4f0b9052f382cade010f3d2757a1c58d5b]
* Authored by Martin Hickey <martin.hickey@ie.ibm.com> on 2016-09-16T15:14:57Z

**Commit 4:**
Update the alert to a warning icon

Fixes #7661

Feedback in review as follows:
For such a large section, I think just displaying a warning icon near the top
of the section would be fine instead of wrapping the entire block in an alert
class.

* Original sha: e3d120df38b5b948589ac9240b1e6cbea984585d [formerly 8be70d4c611091811d48535dcfb924c8be3c5b4b]
* Authored by Martin Hickey <martin.hickey@ie.ibm.com> on 2016-09-19T09:37:14Z

**Commit 5:**
Move the warning message to near end of the page

Fixes #7661

* Original sha: 765cd4c8a669c4c8a59efd867ae92b524ffbbe4f [formerly 14e9bb6c72dd940e4d803d59e33b1393779be29f]
* Authored by Martin Hickey <martin.hickey@ie.ibm.com> on 2016-09-20T16:12:45Z

**Commit 6:**
[indexPatterns/field] show the conflict indices as a list, not JSON

* Original sha: 60b32e2f48002c54bc1dad8b5c2ddf86f223c8ff [formerly 4c96efb6c657205e2b50ac71c6cbacb7cec6f3f6]
* Authored by spalger <email@spalger.com> on 2016-10-13T03:38:15Z
* Committed by Martin Hickey <martin.hickey@ie.ibm.com> on 2016-10-13T10:12:29Z

**Commit 7:**
[indexPatterns/mapping] extract conflict-describing logic

* Original sha: eb6e3cd9ba4e312df7328f6d8fbaf5f22ae93014 [formerly a71dd2c22bd22d885679ba13f378c8ec50338747]
* Authored by spalger <email@spalger.com> on 2016-10-13T03:40:06Z
* Committed by Martin Hickey <martin.hickey@ie.ibm.com> on 2016-10-13T10:13:46Z


Former-commit-id: a407c7e8c5839a485d70ef25eed2a1f8b741d7ff