Backport PR #8691 ---------
**Commit 1:**
Ensure boolean scripted field values are of boolean type

they come back from the server as ‘true’ currently.  Hopefully that
will be fixed in the future, so to future-proof this I am only
converting when the value is not of type boolean.

Fixes #8677

Need to convert both number and bool strings to their respective types.

Throw an error for unexpected types/values.

* Original sha: 278bd926826c585c392cfa609b5eb6cebc0787ce [formerly 7b8ff74513ffeec7fd853ada65848fa651056d2a]
* Authored by Stacey Gammon <gammon@elastic.co> on 2016-10-14T19:50:23Z


Former-commit-id: 7ef0a47069dedfcffc5f4e2725a493da5e91abfc