Backport PR #8758 ---------
**Commit 1:**
URI-encode forwarded location header in proxy

Right now there are situations in which ElasticSearch puts characters in
the character code range between 128 and 255 into the `Location` header.
This leads to an exception when trying to pass on that header through
the hapi proxy in versions before 15.0.0, because it validates that only
US-ASCII characters are used in headers.

To work around that situation, the `Location` header is encoded using
`encodeURI()` for now.

Closes #8705

* Original sha: 03ccb8ec56eb92c86162e307a286305384ba92c3 [formerly 18c23c17d970ad733f4ac74959047fc172cde2d9]
* Authored by Felix Stürmer <stuermer@weltenwort.de> on 2016-10-18T17:55:31Z

**Commit 2:**
Add test to verify umlaut in vis name

Relates to #8705

* Original sha: 238f6f64a55d113f30fd8f61f5330c1c97f23627 [formerly e100e1f5c98fc028846387c12d9d403e9f905758]
* Authored by Felix Stürmer <stuermer@weltenwort.de> on 2016-10-19T09:01:46Z

**Commit 3:**
[elasticsearch/proxy] use different code path with erorr

* Original sha: 0f7e4548c7808d845bf0df661d9a3742b81f14da [formerly fec5e1a2dc0ad772976630c3d128bdae3a6977f3]
* Authored by spalger <email@spalger.com> on 2016-10-19T19:06:39Z


Former-commit-id: afbc7909a2cf516edf10487e399d70e8af7061d7