docs: Note deprecated ES features in breaking changes
Since we introduced the ability for users to manually manipulate some of
the JSON used in Elasticsearch requests in the form of of the filter
editor in 4.3, it's possible that saved objects are taking advantage of
then-deprecated and now-removed Elasticsearch functionality.

This calls out that possibility and links to the Elasticsearch breaking
changes documentation for more information. Even though this is
technically not a Kibana BC break, it manifests as if it were an issue
with Kibana since it results in broken
searches/visualizations/dashboards.

Fixes #8511


Former-commit-id: b9d7d4df6141626146de2cb03ac49221c9f6f746