Tag Cloud should deal with empty responses correctly (#9379)
Manual backport

tag cloud did not work with empty responses correctly. One of the side-effects is that the back button did not work properly. For example, pressing the back button would not empty out the screen and show stale clouds.

This also caused a type-error. Empty configurations meant that we could not access any aggregation-configs to produce a label.

Former-commit-id: 116d9b45dc816a8e8c275c38bddd7b2d981aa55c